PUV updates (#254)
* feat: synth class + write back each function as json result

* fix: Clean up get_vpc_nodes() and its pytest

* feat: Split the main script execution code into each func and add pytest

* feat: Retire json_log and adopt json per rule for both PUV and regular use + linting

* feat: Support QA version and AciVersion instance as input in AciVersion (#259)

* feat: Support QA version and AciVersion instance as input in AciVersion

* fix: Use supported version format in older_than

* fix: Use ValueError in AciVersion

* feat: -c (cversion) / -d (debug_function) args into puv (#265)

* feat: Update synthMaintP with the latest schema with ruleId

* feat: Add decorator `@check_wrapper` for all check functions

A new decorator `@check_wrapper` is to move most of the I/O functionalities,
such as printing and writing the result into a file, outside of each
check function so that each check can focus on the validation logic
itself by minimizing the impact from a requirement change in the output
format and so on.
To support this, a new class `Result` is also introduced to make it
clear what a check function is expected to return. As long as `Result`
class is returned, the decorator `@check_wrapper` handles the printing
them to stdout and files.

* feat: rename synth class to AciResult + breakdown --puv into --api-only and --no-cleanup

* feat: Add --version and --total-checks input args

* feat: Do not touch log folders with some options like --version

* fix: ValueError when col and row len do not match + pytest

* fix: correct checks with mismatched col row length

* fix: cimc logic update for QA

---------

Co-authored-by: takishida <38262981+takishida@users.noreply.github.com>
Co-authored-by: tkishida <tkishida@cisco.com>